+++ src/views/light/LightInstagramVideoCardView.js
@@ -1,12 +1,11 @@
-import CardView from '../CardView.js';
-import InstagramCaptionView from '../InstagramCaptionView.js';
+import InstagramCardView from '../InstagramCardView.js';
 import PlayerOutletView from '../PlayerOutletView.js';
 
-export default class LightInstagramVideoCardView extends CardView {
+export default class LightInstagramVideoCardView extends InstagramCardView {
     constructor() {
         super(...arguments);
 
-        this.instantiates = { InstagramCaptionView, PlayerOutletView };
+        this.instantiates.PlayerOutletView = PlayerOutletView;
         this.template = require('./LightInstagramVideoCardView.html');
     }
 }