+++ src/views/mobile/MobileInstagramImageCardView.js
@@ -1,0 +1,9 @@
+import InstagramCardView from '../InstagramCardView.js';
+
+export default class MobileInstagramImageCardView extends InstagramCardView {
+    constructor() {
+        super(...arguments);
+
+        this.template = require('./MobileInstagramImageCardView.html');
+    }
+}