+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # MiniReel Player 2.0
 
+* *[v0.23.0-rc1]*
+  * [FEATURE]: Use our own design for mobile Instagram cards rather than Instagram's
+  * [FEATURE]: Added support for sponsored Instagram image and video cards
+
 ## v0.22.0 (August 12, 2015)
 * *[v0.22.0-rc4]*
   * [FIX]: Make the playlist truncate extremely long card titles